Refact: 폴더구조 변경 적용 및 로직 분리
